C++Field-class {Rcpp}	R Documentation

Class "C++Field"

Description

Metadata associated with a field of a class exposed through Rcpp modules

Fields

pointer:

external pointer to the internal (C++) object that represents fields

cpp_class:

(demangled) name of the C++ class of the field

read_only:

Is this field read only

class_pointer:

external pointer to the class this field is from.

Methods

No methods defined with class "C++Field" in the signature.

See Also

The fields slot of the C++Class class is a list of C++Field objects

Examples

showClass("C++Field")

[Package Rcpp version 0.11.1 Index]