Module {Rcpp}	R Documentation

Retrieves an Rcpp module

Description

Retrieves an Rcpp module from a dynamic library, usually associated with a package.

Usage

Module(module, PACKAGE = , where = , mustStart =  )

Arguments

module
Name of the module, as declared in the RCPP_MODULE macro internally

PACKAGE
Passed to getNativeSymbolInfo

where
When the module is loaded, S4 classes are defined based on the internal classes. This argument is passed to setClass

mustStart
TODO

Value

An object of class Module collecting functions and classes declared in the module.

[Package Rcpp version 1.0.11 Index]