Rcpp-deprecated {Rcpp}	R Documentation

Deprecated Functions in the Rcpp Package

Description

These functions are provided for compatibility with older versions of the Rcpp package only, and may be removed in future versions.

Details

loadRcppModules calls should now be replaced by loadModule calls, one per Module.

[Package Rcpp version 0.12.18 Index]