Rcpp.plugin.maker {Rcpp}	R Documentation

Facilitating making package plugins

Description

This function helps packages making inline plugins.

Usage

Rcpp.plugin.maker(
  include.before = "", 
  include.after = "", 
  LinkingTo = unique(c(package, "Rcpp")), 
  Depends = unique(c(package, "Rcpp")), 
  Imports = unique(c(package, "Rcpp")), 
  libs = "", 
  Makevars = NULL, 
  Makevars.win = NULL, 
  package = "Rcpp"
)

Arguments

include.before
Code to be included before the Rcpp.h file

include.after
Code to be included after the Rcpp.h file

LinkingTo
Packages to be added to the LinkingTo field

Depends
Packages to be added to the Depends field [deprecated]

Imports
Packages to be added to the Depends field

libs
library flags

Makevars
content for a Makevars file, or NULL

Makevars.win
content for a Makevars.win file, or NULL

package
The package this plugin is for.

Value

A function that is suitable as a plugin. See for example the RcppArmadillo package that uses this to create its inline plugin.

[Package Rcpp version 1.0.8 Index]