LdFlags {Rcpp}	R Documentation

(Deprecated) Rcpp Linker Flags

Description

In previous versions of Rcpp, LdFlags and RcppLdFlags were used to return the required flags and options for the system linker to link to the Rcpp user library. Since we no longer build or ship a user library, these functions now return an empty string.

Usage

LdFlags()
RcppLdFlags()

Value

An empty string.

Author(s)

Dirk Eddelbuettel and Romain Francois

References

Dirk Eddelbuettel and Romain Francois (2011). Rcpp: Seamless R and C++ Integration. Journal of Statistical Software, 40(8), 1-18. URL http://www.jstatsoft.org/v40/i08/ and available as vignette("Rcpp-introduction").

[Package Rcpp version 0.11.1 Index]