asisWeave {Rcpp}	R Documentation

Process pdf vignettes ‘asis’

Description

Simple ‘asis’ Vignette Processor

Usage

asisWeave(file, ...)

asisTangle(file, ..., pattern = "(|[.][^.]*)[.]asis$")

Arguments

file
character Filename to be processed

...
dots Currently ignored

pattern
character A regular expression describing the filename pattern

Details

To pass pre-made pdf vignettes through “as is”, a simple vignette process is added. It is derived from the more feature-complete one in the R.rsp package. To use it, add files named like the pdf file plus an appended .asis with the vignette metainformation and register the vignette processor, see the examples.

Value

The respective filename is returned, invisibly

Author(s)

Henrik Bengtsson for the original versions in package R.rsp, Dirk Eddelbuettel for the shortened ones used here

Examples

# To register this vignette engine in another package, add
# \code{VignetteBuilder: Rcpp} as well as \code{Suggests: Rcpp} to \code{DESCRIPTION}
# which uses the registration this package provides via
## Not run: tools::vignetteEngine("asis", package = pkgname, pattern = "[.](pdf|html)[.]asis$",
                               weave = asisWeave, tangle = asisTangle)
## End(Not run)

# Use a .asis file as in the Rcpp package, for example Rcpp-FAQ.pdf.asis has these lines:
#   %\VignetteIndexEntry{Rcpp-FAQ}
#   %\VignetteKeywords{Rcpp, FAQ, R, Cpp}
#   %\VignettePackage{Rcpp}
#   %\VignetteEncoding{UTF-8}
#   %\VignetteEngine{Rcpp::asis}

[Package Rcpp version 1.1.1 Index]