loadRcppModules {Rcpp}	R Documentation

Loads Rcpp modules on package startup

Description

Function to simplify loading Rcpp modules contained in a package. This function must be called from the .onLoad function of a package. It uses the RcppModules field of the package DESCRIPTION file to query the names of the modules that the package should export, loads each module, and populate each module into the package NAMESPACE.

Usage

loadRcppModules(direct=TRUE)

Arguments

direct
if TRUE the content of the module is exposed in the namespace. Otherwise, the module is exposed.

See Also

populate

[Package Rcpp version 0.11.1 Index]