populate {Rcpp}	R Documentation

Populates a namespace or an environment with the content of a module

Description

Populates a namespace or an environment with the content of a module

Usage

populate(module, env)

Arguments

module
Rcpp module

env
environment or namespace

[Package Rcpp version 0.11.1 Index]