C++Constructor-class {Rcpp}	R Documentation

Class "C++Constructor"

Description

Representation of a C++ constructor

Extends

Class "envRefClass", directly. Class ".environment", by class "envRefClass", distance 2. Class "refClass", by class "envRefClass", distance 2. Class "environment", by class "envRefClass", distance 3, with explicit coerce. Class "refObject", by class "envRefClass", distance 3.

Fields

pointer:

pointer to the internal structure that represent the constructor

class_pointer:

pointer to the internal structure that represent the associated C++ class

nargs:

Number of arguments the constructor expects

signature:

C++ signature of the constructor

docstring:

Short description of the constructor

[Package Rcpp version 1.0.8 Index]