C++Object-class {Rcpp}	R Documentation

c++ internal objects

Description

C++ internal objects instanciated from a class exposed in an Rcpp module

Objects from the Class

This is a virtual class. Actual C++ classes are subclasses.

Methods

$

signature(x = "C++Object"): invokes a method on the object, or retrieves the value of a property

$<-

signature(x = "C++Object"): set the value of a property

show

signature(object = "C++Object"): print the object

[Package Rcpp version 1.0.8 Index]