C++OverloadedMethods-class {Rcpp}	R Documentation

Class "C++OverloadedMethods"

Description

Set of C++ methods

Extends

Class "envRefClass", directly. Class ".environment", by class "envRefClass", distance 2. Class "refClass", by class "envRefClass", distance 2. Class "environment", by class "envRefClass", distance 3, with explicit coerce. Class "refObject", by class "envRefClass", distance 3.

Fields

pointer:

Object of class externalptr pointer to the internal structure that represents the set of methods

class_pointer:

Object of class externalptr pointer to the internal structure that models the related class

[Package Rcpp version 1.0.8 Index]