Rcpp-deprecated {Rcpp}	R Documentation

Deprecated Functions in the Rcpp Package

Description

These functions are provided for compatibility with older versions of the Rcpp package only, and may be removed in future versions.

Details

loadRcppModules calls should now be replaced by loadModule calls, one per Module.

LdFlags and RcppLdFlags are no longer required as no library is provided (or needed) by Rcpp (as it was up until release 0.10.1).

Author(s)

Dirk Eddelbuettel and Romain Francois

[Package Rcpp version 1.0.11 Index]