Rcpp-internal {Rcpp}	R Documentation

Rcpp internal functions

Description

The Rcpp utilizes several internal functions which need to visible within its namespace, yet are not user-facing. This page provides basic documentation for them.

Usage

  cpp_object_initializer(.self, .refClassDef, ..., .object_pointer)
  cpp_object_dummy(.self, .refClassDef)

Arguments

.self
A reference to the object being modified

.refClassDef
The definition of a reference class

...
Further arguments

.object_pointer
Optional external pointer

[Package Rcpp version 1.0.11 Index]