Rcpp-package {Rcpp}	R Documentation

R / C++ interface

Description

The Rcpp package provides C++ classes that greatly facilitate interfacing C or C++ code in R packages using the .Call interface provided by R.

Introduction

Rcpp provides C++ classes to facilitate manipulation of a large number of R data structures : vectors, functions, environments, ...

The “Rcpp-introduction” vignette gives an introduction on the package

Usage for package building

The “Rcpp-package” vignette documents how to use Rcpp in client packages.

History

The initial versions of Rcpp were written by Dominick Samperi during 2005 and 2006.

Dirk Eddelbuettel made some additions, and became maintainer in 2008.

Dirk Eddelbuettel and Romain Francois have been extending Rcpp since 2009.

Author(s)

Dirk Eddelbuettel and Romain Francois

References

Dirk Eddelbuettel and Romain Francois (2011). Rcpp: Seamless R and C++ Integration. Journal of Statistical Software, 40(8), 1-18, doi: 10.18637/jss.v040.i08. Also available as vignette("Rcpp-introduction").

Eddelbuettel, Dirk (2013) Seamless R and C++ Integration with Rcpp. Springer, New York. ISBN 978-1-4614-6867-7.

See Also

Development for Rcpp can be followed via the GitHub repository at https://github.com/RcppCore/Rcpp.

Extensive examples with full documentation are available at https://gallery.rcpp.org.

Examples

## Not run: 
# introduction to Rcpp
vignette("Rcpp-introduction")

# information on how to build a package that uses Rcpp
vignette("Rcpp-package")

## End(Not run)

[Package Rcpp version 1.0.8 Index]