LdFlags-deprecated {Rcpp}	R Documentation

Deprecated Rcpp Linker Flags

Description

In Rcpp versions prior to release 0.10.1 of November 2013, LdFlags and RcppLdFlags were used to return the required flags and options for the system linker to link to the Rcpp user library. Since we no longer build or ship a user library, these functions now return an empty string. As of Rcpp release 0.12.19, these functions are now deprecated.

Usage

  LdFlags()
  RcppLdFlags()

Value

An empty string.

Author(s)

Dirk Eddelbuettel and Romain Francois

References

Dirk Eddelbuettel and Romain Francois (2011). Rcpp: Seamless R and C++ Integration. Journal of Statistical Software, 40(8), 1-18, doi:10.18637/jss.v040.i08. Also available as vignette("Rcpp-introduction").

[Package Rcpp version 1.0.11 Index]