RcppUnitTests {Rcpp}	R Documentation

Rcpp : unit tests results

Description

Unit tests results for package Rcpp.

Unit tests are run automatically at build time and reports are included in the doc directory as html or text.

See Also

../doc/unitTests-results/Rcpp-unitTests.html: html formatted unit test report

Examples

# unit tests are in the unitTests directory of the package
list.files( system.file("unitTests", package = "Rcpp" ),
	pattern = "^runit", full = TRUE )

# trigger the unit tests preparation, follow printed instructions
# on how to run them
## Not run: 
source( system.file("unitTests", "runTests.R", package = "Rcpp" ) )

## End(Not run)

[Package Rcpp version 1.0.8 Index]