compilerCheck {Rcpp}	R Documentation

Check for Minimal (g++) Compiler Version

Description

Helper function to establish minimal compiler versions, currently limited only to g++ which (particularly for older RHEL/CentOS releases) is too far behind current C++11 standards required for some packages.

Usage

compilerCheck(minVersion = package_version("4.6.0"))

Arguments

minVersion
An object of type package_version, with a default of version 4.6.0

Details

This function looks up g++ (as well as optional values in the CXX and CXX1X environment variables) in the PATH. For all values found, the output of g++ -v is analyzed for the version string, which is then compared to the given minimal version.

Value

A boolean value is returned, indicating if the minimal version is being met

Author(s)

Dirk Eddelbuettel

[Package Rcpp version 1.0.8 Index]