getRcppVersion {Rcpp}	R Documentation

Export the Rcpp (API) Package Version

Description

Helper function to report the package version of the R installation.

Usage

getRcppVersion(devel = FALSE)

Arguments

devel
An logical value indicating if the development or release version number should be returned, default is release.

Details

While packageVersion(Rcpp) exports the version registers in DESCRIPTION, this version does get incremented more easily during development and can therefore be higher than the released version. The actual #define long used at the C++ level corresponds more to an ‘API Version’ which is now provided by this function, and use for example in the package skeleton generator.

Value

A package_version object with either the release or development version.

Author(s)

Dirk Eddelbuettel

See Also

packageVersion, Rcpp.package.skeleton

Examples

getRcppVersion()

[Package Rcpp version 1.0.8 Index]