registerPlugin {Rcpp}	R Documentation

Register an inline plugin

Description

Register an inline plugin for use with sourceCpp or cppFunction. Inline plugins are functions that return a list with additional includes, environment variables, and other compilation context.

Usage

registerPlugin(name, plugin)

Arguments

name
Name of the inline plugin

plugin
Inline plugin function

Details

Plugins can be added to sourceCpp compilations using the Rcpp::plugins attribute.

See Also

Rcpp::plugins

[Package Rcpp version 1.0.11 Index]